:nail_care: Reorganization for utils methods
