+++ src/gitHandler/index.js
@@ -9,7 +9,7 @@ async function userValue(prop) {
   let data = '';
 
   try {
-    data = await utils.execp(`git config user.${prop}`);
+    data = await utils.process.execp(`git config user.${prop}`);
   } catch (error) {
     console.error(error);
   }
@@ -26,7 +26,7 @@ async function init(path = '.') {
   let resp = '';
 
   try {
-    resp = await utils.execp('git init', path);
+    resp = await utils.process.execp('git init', path);
   } catch (error) {
     console.error(error);
   }
@@ -44,7 +44,7 @@ async function commit(path = '.', msg = 'Initial commit') {
   let resp = '';
 
   try {
-    resp = await utils.execp(`git add . && git commit -m'${msg}'`, path);
+    resp = await utils.process.execp(`git add . && git commit -m'${msg}'`, path);
   } catch (error) {
     console.error(error);
   }
@@ -63,7 +63,7 @@ async function addRemote(path = '.', url, remote = 'origin') {
   let resp = '';
 
   try {
-    resp = await utils.execp(`git remote add ${remote} ${url}`, path);
+    resp = await utils.process.execp(`git remote add ${remote} ${url}`, path);
   } catch (error) {
     console.error(error);
   }
@@ -82,7 +82,7 @@ async function push(path = '.', remote = 'origin', branch = 'master') {
   let resp = '';
 
   try {
-    resp = await utils.execp(`git push ${remote} ${branch}`, path);
+    resp = await utils.process.execp(`git push ${remote} ${branch}`, path);
   } catch (error) {
     console.error(error);
   }
