+++ src/githubHandler/index.js
@@ -27,7 +27,7 @@ async function create(name, isPrivate = false, description = '', url = '') {
   const cmd = `curl -w "%{http_code}" -H "Authorization: token ${token}" -d '{"name": "${name}", "private": ${isPrivate}, "description": "${description}", "homepage": "${url}"}' https://api.github.com/user/repos`;
 
   try {
-    result = await utils.execp(cmd);
+    result = await utils.process.execp(cmd);
   } catch (error) {
     console.error(error);
     return false;
@@ -74,7 +74,7 @@ async function deleteRepo(name, user) {
   const cmd = `curl -w "%{http_code}" -XDELETE -H "Authorization: token ${token}" https://api.github.com/repos/${user}/${name}`;
 
   try {
-    result = await utils.execp(cmd);
+    result = await utils.process.execp(cmd);
   } catch (error) {
     console.error(error);
     return false;
