+++ test/cleanup.js
@@ -2,11 +2,11 @@ const utils = require('../src/utils');
 const githubHandler = require('../src/githubHandler');
 
 function cleanup(path = 'a-demo-project') {
-  const name = utils.normalizeName(path);
+  const name = utils.string.normalizeName(path);
 
   // Remove test folder
   console.log(`Deleting folder ${path}`);
-  utils.deleteDirRecursive(path);
+  utils.fs.deleteDirRecursive(path);
 
   // Delete github project
   githubHandler.deleteRepo(name, 'nmicht');
